Allow non centered tic labels for time ranges
While '7.Apr.2003 18:45:20' is an instant in time "May 2014" or "Week 15 (2015)" is more a
period. If tics are normaly labeled with "May 2104" it is not absolutely clear that this means
"1st of May 2014 00:00:00".
This commit allows to left justify labels on date/time axis if the tic label describes more
a time span than an time instant. Like this a label like "May 2014" lies completely in the
range of may 2014 which makes it a bit clearer. The change from centered labels to left justified
happens when no more time is displayed.